requirements.txt
# StudyAI AI Engine - Requirements

# Core Framework
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0

# AI and ML
openai==1.3.7
langchain==0.0.339
langchain-community==0.0.10
langchain-openai==0.0.2

# Vector Database (choose one)
pinecone-client==2.4.0
chromadb==0.4.18

# Database and Caching
sqlalchemy==2.0.23
alembic==1.13.0
redis==5.0.1
psycopg2-binary==2.9.9

# Utilities
python-dotenv==1.0.0
loguru==0.7.2
httpx==0.25.2
tenacity==8.2.3

# Development
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
flake8==6.1.0
mypy==1.7.1

# Educational Content Processing
numpy==1.25.2
pandas==2.1.4
scikit-learn==1.3.2
matplotlib==3.8.2

# Optional: For future mathematical processing
sympy==1.12
scipy==1.11.4